+++ source/tasks/Publish_GitHub_Wiki_Content.build.ps1
@@ -117,60 +117,8 @@ task Publish_GitHub_Wiki_Content {
             $local:DebugPreference = 'Continue'
         }
 
-        $OutputDirectory = Get-SamplerAbsolutePath -Path $OutputDirectory -RelativeTo $BuildRoot
-
-        "`tOutputDirectory       = '$OutputDirectory'"
-
-        $BuiltModuleSubdirectory = Get-SamplerAbsolutePath -Path $BuiltModuleSubdirectory -RelativeTo $OutputDirectory
-
-        if ($VersionedOutputDirectory)
-        {
-            <#
-                VersionedOutputDirectory is not [bool]'' nor $false nor [bool]$null
-                Assume true, wherever it was set
-            #>
-            $VersionedOutputDirectory = $true
-        }
-        else
-        {
-            <#
-                VersionedOutputDirectory may be [bool]'' but we can't tell where it's
-                coming from, so assume the build info (Build.yaml) is right
-            #>
-            $VersionedOutputDirectory = $BuildInfo['VersionedOutputDirectory']
-        }
-
-        $GetBuiltModuleManifestParams = @{
-            OutputDirectory          = $OutputDirectory
-            BuiltModuleSubdirectory  = $BuiltModuleSubDirectory
-            ModuleName               = $ProjectName
-            VersionedOutputDirectory = $VersionedOutputDirectory
-            ErrorAction              = 'Stop'
-        }
-
-        $builtModuleManifest = Get-SamplerBuiltModuleManifest @GetBuiltModuleManifestParams
-        $builtModuleManifest = [string](Get-Item -Path $builtModuleManifest).FullName
-
-        "`tBuilt Module Manifest         = '$builtModuleManifest'"
-
-        $builtModuleBase = Get-SamplerBuiltModuleBase @GetBuiltModuleManifestParams
-        $builtModuleBase = [string](Get-Item -Path $builtModuleBase).FullName
-
-        "`tBuilt Module Base             = '$builtModuleBase'"
-
-        $moduleVersion = Get-BuiltModuleVersion @GetBuiltModuleManifestParams
-        $moduleVersionObject = Split-ModuleVersion -ModuleVersion $moduleVersion
-        $moduleVersionFolder = $moduleVersionObject.Version
-        $preReleaseTag = $moduleVersionObject.PreReleaseString
-
-        "`tModule Version                = '$ModuleVersion'"
-        "`tModule Version Folder         = '$moduleVersionFolder'"
-        "`tPre-release Tag               = '$preReleaseTag'"
-
-        "`tProject Path                  = $ProjectPath"
-        "`tProject Name                  = $ProjectName"
-        "`tSource Path                   = $SourcePath"
-        "`tBuilt Module Base             = $builtModuleBase"
+        # Get the values for task variables
+        . Set-SamplerTaskVariable
 
         # If variables are not set then update variables from the property values in the build.yaml.
         foreach ($gitHubConfigKey in @('GitHubConfigUserName', 'GitHubConfigUserEmail'))