+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix missing verbose message in function `Invoke-Git`.
 - Fix variable reference for localized strings so they passed HQRM tests.
 - Fix statement so function passed HQRM tests.
+- Tasks now uses `Set-SamplerTaskVariable` from the module Sampler to set
+  the common build task variables.
 
 ## [0.10.3] - 2022-01-26